# manual/static/openapi/admin-api-openapi.json
{
  "openapi": "3.0.2",
...
  "info": {
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "contact": {
      "name": "MAIF open source team",
      "email": "oss@maif.fr"
    },
    "description": "Admin API of Daikoku",
    "title": "Daikoku Admin API",
    "version": "16.3.0"
  },
...
  "components": {
    "schemas": {
      "Tenant": {
        "description": "Tenant",
        "properties": {
          "_id": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "_deleted": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string",
            "nullable": true
          },
          "contact": {
            "type": "string"
          },
          "style": {
            "$ref": "#/components/schemas/DaikokuStyle",
            "nullable": true
          },
          "defaultLanguage": {
            "type": "string",
            "nullable": true
          },
          "otoroshiSettings": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/OtoroshiSettings"
            }
          },
          "mailerSettings": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/MailerSettings.ConsoleMailerSettings"
              },
              {
                "$ref": "#/components/schemas/MailerSettings.MailgunSettings"
              },
              {
                "$ref": "#/components/schemas/MailerSettings.MailjetSettings"
              },
              {
                "$ref": "#/components/schemas/MailerSettings.SimpleSMTPSettings"
              },
              {
                "$ref": "#/components/schemas/MailerSettings.SendgridSettings"
              }
            ]
          },
          "bucketSettings": {
            "nullable": true,
            "$ref": "#/components/schemas/S3Configuration"
          },
          "authProvider": {
            "nullable": true,
            "type": "string",
            "enum": [
              "local",
              "otoroshi",
              "ldap",
              "oauth2"
            ]
          },
          "authProviderSettings": {
            "type": "object"
          },
          "auditTrailConfig": {
            "nullable": false,
            "$ref": "#/components/schemas/AuditTrailConfig"
          },
          "isPrivate": {
            "type": "boolean",
            "nullable": true
          },
          "adminApi": {
            "type": "string"
          },
          "adminSubscriptions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "creationSecurity": {
            "type": "boolean",
            "nullable": true
          },
          "subscriptionSecurity": {
            "type": "boolean",
            "nullable": true
          },
          "apiReferenceHideForGuest": {
            "type": "boolean",
            "nullable": true
          },
          "defaultMessage": {
            "type": "string",
            "nullable": true
          },
          "tenantMode": {
            "type": "string",
            "enum": [
              "default",
              "maintenance",
              "construction",
              "translation"
            ],
            "nullable": true
          },
          "aggregationApiKeysSecurity": {
            "type": "boolean",
            "nullable": true
          },
          "robotTxt": {
            "type": "string",
            "nullable": true
          },
          "thirdPartyPaymentSettings": {
            "type": "array",
            "nullable": true,
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ThirdPartyPaymentSettings.StripeSettings"
                }
              ]
            }
          },
          "display": {
            "nullable": true,
            "type": "string",
            "enum": [
              "default",
              "environment"
            ]
          },
          "environments": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "_id",
          "enabled",
          "name",
          "contact",
          "adminApi"
        ]
      },
...
      "done": {
        "description": "task is done",
        "properties": {
          "done": {
            "type": "boolean"
          }
        }
      },
      "error": {
        "description": "error response",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "patch": {
        "description": "A set of changes described in JSON Patch format: http://jsonpatch.com/ (RFC 6902)",
        "type": "array",
        "items": {
          "type": "object",
          "required": [
            "op",
            "path"
          ],
          "properties": {
            "op": {
              "type": "string",
              "enum": [
                "add",
                "replace",
                "remove",
                "copy",
                "test"
              ]
            },
            "path": {
              "type": "string"
            },
            "value": {}
          }
        }
      },
...
      "User": {
        "description": "User",
        "properties": {
          "_id": {
            "type": "string"
          },
          "_deleted": {
            "type": "boolean"
          },
          "tenants": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "origins": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string",
              "enum": [
                "local",
                "otoroshi",
                "ldap",
                "oauth2"
              ]
            }
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "picture": {
            "type": "string",
            "nullable": true
          },
          "pictureFromProvider": {
            "type": "boolean",
            "nullable": true
          },
          "personalToken": {
            "type": "string",
            "nullable": true
          },
          "isDaikokuAdmin": {
            "type": "boolean",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true,
            "description": "optionnal value but required in case of tenant has a Local AuthProvider"
          },
          "hardwareKeyRegistrations": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object"
            }
          },
          "lastTenant": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "defaultLanguage": {
            "type": "string",
            "nullable": true
          },
          "starredApis": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "twoFactorAuthentication": {
            "$ref": "#/components/schemas/TwoFactorAuthentication",
            "nullable": true
          },
          "invitation": {
            "$ref": "#/components/schemas/UserInvitation",
            "nullable": true
          }
        },
        "required": [
          "_id",
          "name",
          "email"
        ]
      },
...
    }
  },
  "paths": {
    "/admin-api/tenants/{id}": {
      "delete": {
        "summary": "delete a tenant",
        "operationId": "tenants.delete",
        "responses": {
          "200": {
            "description": "entity deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/done"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The ID of the tenant to delete",
            "required": true
          }
        ],
        "tags": [
          "tenant"
        ]
      },
      "patch": {
        "summary": "update a tenant with JSON patch or by  merging JSON object",
        "operationId": "tenants.patch",
        "requestBody": {
          "description": "the patch to update the tenant or a JSON object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/patch"
                  },
                  {
                    "$ref": "#/components/schemas/Tenant"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Tenant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The tenant ID to update",
            "required": true
          }
        ],
        "tags": [
          "tenant"
        ]
      },
      "put": {
        "summary": "update a tenant",
        "operationId": "tenants.update",
        "requestBody": {
          "description": "the ID of the tenant to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tenant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Tenant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The ID of the tenant to update",
            "required": true
          }
        ],
        "tags": [
          "tenant"
        ]
      },
      "get": {
        "summary": "read a tenant",
        "operationId": "tenants.findById",
        "responses": {
          "200": {
            "description": "found entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Tenant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The ID of the tenant to retrieve",
            "required": true
          }
        ],
        "tags": [
          "tenant"
        ]
      }
    },
...
    "/admin-api/users/{id}": {
      "delete": {
        "summary": "delete a user",
        "operationId": "users.delete",
        "responses": {
          "200": {
            "description": "entity deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/done"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "patch": {
        "summary": "update a user with JSON patch or by  merging JSON object",
        "operationId": "users.patch",
        "requestBody": {
          "description": "the patch to update the user or a JSON object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/patch"
                  },
                  {
                    "$ref": "#/components/schemas/User"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "put": {
        "summary": "update a user",
        "operationId": "users.update",
        "requestBody": {
          "description": "the user to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "get": {
        "summary": "read a user",
        "operationId": "users.findById",
        "responses": {
          "200": {
            "description": "found entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
...
    "/admin-api/teams/{id}": {
      "delete": {
        "summary": "delete a team",
        "operationId": "teams.delete",
        "responses": {
          "200": {
            "description": "entity deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/done"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "team"
        ]
      },
      "patch": {
        "summary": "update a team with JSON patch or by  merging JSON object",
        "operationId": "teams.patch",
        "requestBody": {
          "description": "the patch to update the team or a JSON object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/patch"
                  },
                  {
                    "$ref": "#/components/schemas/Team"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "team"
        ]
      },
      "put": {
        "summary": "update a team",
        "operationId": "teams.update",
        "requestBody": {
          "description": "the team to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Team"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "team"
        ]
      },
      "get": {
        "summary": "read a team",
        "operationId": "teams.findById",
        "responses": {
          "200": {
            "description": "found entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "team"
        ]
      }
    },
...
  }
}